Further cleanups to directory structure etc
Still need to resolve the warnings from glib-gettextize.

By using (non-glib)gettextize and the accompanying AC-macros, this seems to
work but introduces a world of pain in annoying build artefacts. Some of it
is apparently possible to work around (like automated ChangeLog updates)
but this needs to be done properly.