Attempt to update AUTHORS
This list is probably incomplete, I regret to say. I am also trying to
balance it more. The Beryl-team was largely left as minor contributors,
which is not Right™. The projects merged, the work on Beryl is still very
much relevant.

I will continue to update this list as I merge other components.