Start flipping the dependency-check logic
It's up-side down, people.

If ten plugins need a package, we are defining 10 individual namespaces for
the same thing. (E.g. ANNOTATE_CFLAGS instead of CAIRO_XLIB_XRENDER_CFLAGS).
This makes the makefiles more ambiguous (CAIRO_CFLAGS is more informative
than ANNOTATE_CFLAGS), gets us to check for the same things multiple times,
and is generally meh.

This patch fixes this for annotate and adds optional building of wall.

The wall-logic should also be improved since it's checking for
cairo-xlib-xrender (same as annotate) in addition to cairo. I don't know if
they actually need the same things or not at the moment.

This patch also begins the long road of populating the new
'missing_optional_deps' variable which will help users see what
libraries/components are missing on their components, but optional. It needs
to be populated by basically every optional dependency check.