Continue flipping configure.ac logic
I am also renaming a number of variables from FOO_PLUGIN to USE_FOO, which
is arguably not perfect. The reason for this is that there was a mix, and
it'll be easier to unify first then, potentially, change it back.

Incidentally this'll make things like "these three plugins all require DBUS"
easier to implement, so maybe it's not such a bad thing.

The missing_optional_deps is almost useful already. Feedback wanted though.