Re-introduced dbus-glib for gtk-window-decorator
I checked src/ and plugins/ before nuking it and discovered it was obsolete
in the process. Alas, gtk-window-decorator still uses it. And unlike most of
the other AC_DEFINE stuff that's present in configure.ac, this one is
actually needed.

At some point we should fix the dbus-glib dep though, what with the
obsolescence and all.