Fix numerous shadowed variables
Guys, wtf.

Seriously.

This is why you don't write scripts in C.

Most of these are harmless, but most are also idiotic. Example:

In rotate, why would you need to get an option for every variation of
if-clauses, and even REDEFINE the int storing the option? Do you actually
think it changed between the top of the function and a little further down?

gtk-window-decorator is actually quite confusing. It has a shitload of
global variables, and then shadows one of them. I didn't even know if this
was a redundant variable, or if it might differ from the global one. I ended
up doing a cosmetic fix, but adding a comment explaining it.

And let's not even talk about the shit in core. Or how people manage to
shadow int w to CompWindow *w and not notice (THAT SHOULD TELL YOU SOMETHING
ABOUT THE SIZE OF YOUR FUNCTION).

This is what happens when you write functions that are 123124125 lines long
and needs an 8MB stack to keep all the local variables.