Tweak debug output and add simple log-macros
compLogMessage isn't used and isn't fine grained enough for people to debug.
compLog(), compWarn() and compDebug() are all trivial to use, and reasonably
self-explanatory.

The output might be somewhat intimidating to users, but this is mostly going
to end up in ~/.xsession-errors, and it's better to have useful messages
than no messages.

I'll be using this and gradually removing compLogMessage() over time.