Add some verbosity to image loading
This alleviates a bit of #3 , but isn't a good solution.

The fundamental problem is listed in #27 :

Image plugins are wacko, and the "try every combination of paths and
suffixes we can imagine" thing is making it much more complex than it
should be to report why we can't open an image.