Improve image loading path expansion logic
This is the first step in moving away from abstract file names and shifting
responsibility of FINDING the file away from the plugins that LOAD the file.

The next step is to remove suffix-stuff from the plugins themself, and
removing the now redundant "path" argument to the hook.

Oh, and I also fixed so core.xml.in gets built. Oops.