Simplify path-logic for image handling
Fixes #27
Fixes #3

This removes the concept of dynamic paths from image format plugins.
Responsibility for finding a file is shifted to core, and the plugins just
have to try to parse the file.

It also adds a decent amount of error handling output for the user to solve
problems, but without flooding the terminal. Errors for specific paths are
only printed if the error is NOT "File not found". This means that
permission-problems will be obvious, and we will not continue to look for a
file if we found one that we couldn't read.