Remove release-automation
Probably not doing it like that in the future, if at all. Seems misleading
to keep it.

Bear in mind that this does not make the information in those files less
useful, but I'd rather go back in git to view how things was done in the
past than have to deal with a mix of current and out-of-date information.